+++ client/tws_foundation_client/lib/src/services/business/trailers_externals_service.dart
@@ -28,5 +28,11 @@ final class TrailersExternalsService extends TrailersExternalsServiceBase {
     CSMActEffect actEffect = await twsPost('update', trailer, auth: auth);
     return MainResolver<RecordUpdateOut<TrailerExternal>>(actEffect);
   }
+
+  @override
+  Effect<TrailerExternal> delete(TrailerExternal trailer, String auth) async {
+    CSMActEffect actEffect = await twsPost('delete', trailer, auth: auth);
+    return MainResolver<TrailerExternal>(actEffect);
+  }
 }