+++ server/TWS Foundation/Server.Quality/Suit/Controllers/Security/Q_AccountsService.cs
@@ -194,6 +194,7 @@ public class Q_AccountsService
         }
         #endregion
     }
+
     [Fact]
     public async Task Delete() {
         string testTag = Guid.NewGuid().ToString()[..4];