+++ server/TWS Foundation/TWS Foundation/Controllers/Business/TrailersExternalsController.cs
@@ -22,13 +22,18 @@ public class TrailersExternalsController : ControllerBase {
         return Ok(await Service.View(Options));
     }
 
-    [HttpPost(), Auth("TrailersExternals","Create")]
+    [HttpPost(), Auth("TrailersExternals", "Create")]
     public async Task<IActionResult> Create(TrailerExternal[] Trailers) {
         return Ok(await Service.Create(Trailers));
     }
 
-    [HttpPost(), Auth("TrailersExternals","Update")]
+    [HttpPost(), Auth("TrailersExternals", "Update")]
     public async Task<IActionResult> Update(TrailerExternal Trailer) {
         return Ok(await Service.Update(Trailer));
+    }
+
+    [HttpPost(), Auth("TrailersExternals", "Delete")]
+    public async Task<IActionResult> Delete(TrailerExternal Trailer) {
+        return Ok(await Service.Delete(Trailer));
     }
 }